requirements.txt
streamlit
fastapi
uvicorn
google-cloud-storage
google-auth
requests